Sorting into b working, tested with 10 numbers. High to low on B, Low to High on A.

+++ sources/lists.c
@@ -6,7 +6,7 @@
 /*   By: owen <owen@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/08 12:06:09 by owen          #+#    #+#                 */
-/*   Updated: 2025/02/13 15:49:34 by owen          ########   odam.nl         */
+/*   Updated: 2025/02/18 15:47:33 by owhearn       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-size_t	ps_lst_size(t_stack *list)
+int	ps_lst_size(t_stack *list)
 {
-	size_t	size;
+	int		size;
 	t_stack	*temp;
 
 	size = 0;
